Add links to LSHTM and twitter where available
Link each author to their page on LSHTM, where one could be found, and
also to twitter where they have a profile.

I've gathered this information from the LSHTM website, but it's
possible that authors have twitter accounts that are not linked there,
or other URLs instead of LSHTM ones.