Use tags to navigate posts, rather than categories
It's not super-clean, but posts and pages can have multiple tags
without it affecting the URL structure of the site, unlike categories,
so if we want posts to appear in multiple sections, we basically have
no choice but to use them.

This also revises the navigation to a vertical set of pills on the
left hand side, most auto-generated and some hard-coded.

I've moved the page title from the default layout into some of the
specific layouts to allow the title to be optionally inferred from the
tag filter name on category pages, and for that title to appear over
the content, rather than over the whole page (i.e. on top of the nav).