Unify translation mechanisms into one
Now that we're not using the category information as data, and instead
are using the post tags, we don't actually need the categories data
file, except for its translations. So, instead, let's unify the
translation mechanisms into a single one, and consistently infer the
target language from the page, since it's always set.