Simplify translated posts workflow
Now that all posts have a language, we don't actually need to keep
translated ones in a separate directory, and they can live alongside
their original versions, sharing the same figures and so on. It's only
important that they have the language set correctly, and then they
will be automatically filtered for us.