Flatten post structure now that categories aren't used
This should make it simpler for people to add new posts, since they
don't need to decide which category of several the post might appear
in.

The downside is that it's more possible that people will try to put
generic named figures and reports into the shared directories, but
this was already a bit of a problem anyway.