Add new tags for post categorisation
Once at least one post exists with the new tag, it will appear in the
left-hand navigation.

Once all posts have been moved to the new tags, the old categories
will disappear, and their topic pages can be removed. If a tag is a
"renaming" of an old tag, then the old url for the old topic page
should be added to the `redirect_from` front-matter of the new tag.